+++ batch/src/main/java/com/comeeatme/batch/address/AddressCodeDto.java
@@ -1,4 +1,4 @@
-package com.comeeatme.batch.job.address;
+package com.comeeatme.batch.address;
 
 import lombok.Value;
