+++ batch/src/main/java/com/comeeatme/batch/restaurant/RestaurantSkipListener.java
@@ -1,12 +1,11 @@
-package com.comeeatme.batch.job.restaurant.init;
+package com.comeeatme.batch.restaurant;
 
-import com.comeeatme.batch.job.restaurant.LocalDataRestaurantDto;
 import com.comeeatme.domain.restaurant.LocalData;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.SkipListener;
 
 @Slf4j
-public class RestaurantInitSkipListener implements SkipListener<LocalDataRestaurantDto, LocalData> {
+public class RestaurantSkipListener implements SkipListener<LocalDataRestaurantDto, LocalData> {
 
     @Override
     public void onSkipInRead(Throwable t) {
