+++ batch/src/main/java/com/comeeatme/batch/restaurant/RestaurantTrimProcessor.java
@@ -1,6 +1,6 @@
-package com.comeeatme.batch.job.restaurant.init;
+package com.comeeatme.batch.restaurant;
 
-import com.comeeatme.batch.job.restaurant.LocalDataRestaurantDto;
+import com.comeeatme.batch.restaurant.LocalDataRestaurantDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.util.StringUtils;
