+++ batch/src/main/java/com/comeeatme/batch/restaurant/update/RestaurantUpdateFileDownloadJobConfig.java
@@ -1,7 +1,7 @@
-package com.comeeatme.batch.job.restaurant.update;
+package com.comeeatme.batch.restaurant.update;
 
 
-import com.comeeatme.batch.job.restaurant.RestaurantFileConstant;
+import com.comeeatme.batch.restaurant.RestaurantFileConstant;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.Job;
